Moved base.py to app/service/auth -  Updated imports in tests and related modules
