+++ margin_app/app/api/auth.py
@@ -10,7 +10,7 @@ from loguru import logger
 from pydantic import EmailStr
 
 from app.core.config import settings
-from app.auth.base import google_auth
+from app.services.auth.base import google_auth
 from app.crud.admin import admin_crud
 from app.schemas.admin import AdminResetPassword
 from app.services.auth import (
