+++ margin_app/app/main.py
@@ -17,7 +17,7 @@ from app.api.order import router as order_router
 from app.api.pools import router as pool_router
 from app.api.user import router as user_router
 from app.api.auth import router as auth_router
-from app.auth.base import get_current_user
+from app.services.auth.base import get_current_user
 from app.models.admin import Admin
 
 # Initialize FastAPI app
