+++ margin_app/app/services/emails.py
@@ -4,7 +4,7 @@ from typing import Optional, Dict, Any
 from sendgrid import SendGridAPIClient
 from sendgrid.helpers.mail import Mail, Email, To, Content, TemplateId
 from app.core.config import settings
-from app.auth.base import create_access_token, get_expire_time
+from app.services.auth.base import create_access_token, get_expire_time
 import logging
 
 logger = logging.getLogger("EmailService")
