+++ margin_app/app/tests/test_auth_admin_user_middleware.py
@@ -5,7 +5,7 @@ Testing module for auth admin user middleware (app.main.auth_admin_user).
 import pytest
 from fastapi import status, FastAPI, Request
 
-from app.auth.base import create_access_token
+from app.services.auth.base import create_access_token
 from app.models.admin import Admin
 from app.crud.admin import admin_crud
 from app.tests.api.test_admin_api import (
